telegram_english_bot/requirements.txt
python-telegram-bot==13.15
urllib3==1.26.6
six==1.16.0